requirements.txt
flask
gunicorn
Flask
flask-cors
google-api-python-client
google-auth
google-auth-httplib2
google-auth-oauthlib
gspread
gspread-formatting
oauth2client
python-dotenv